Fixed "ButtonBasic" hover on mobile devices
Hover on mobile devices looked ugly.
At first I tried to solve it only with css, adding animation
of pressing to button while it is :active.
But this didn't seem to work with button type=submit, so I came up
with adding class "clicked" each time user clicks button and
removing it after 0.5s when animation is over.