Added "React-slick" slider to "Blog" page
To add incredibly beautiful :) animation to slick arrows on mobile
devices I made "SlickArrowsMobileAnimation" class in "helpers.js".
This class accepts className of slider and time that is needed to
perform animation.
After we make an instance of that class it gets arrows of certain
slider and listens them, when they are clicked "clicked" className
adds and removes after animationTime passed to constructor is over.
Also the class has a method to remove listeners and timers if
component will unmount.
Basically class adds to slick arrows "clicked" class name when
they're clicked and in css we can add animation for hover
devices when "clicked" class is applied.

Also I corrected widths and margins and paddings of some elements on
a "Blog" page.