"Pagination" update & new "usePagination" hook
I extracted pagination logic from "Blogs" and "MyBlogs" pages into
"usePagination" hook. Also in hook I save current page to local
storage and when page refreshes user still is on the same page.
I added variables that store names of local storage variables in
"constants.ts". Each page stores value for pagination in its
own variable.

I added to BlogResponseI optional "fake" key. It's needed
since I get blogs from server on "Blogs" page and add to them fake
blogs to show how my pagination works.
I added fake key since every blog that's not fake requests author
from server and this can eat my free requests to database.

"helpers.ts" in pagination folder changed a lot.
Now if we have lots of pages, pagination will show link not only
to the last page, but also to the first one.
If we are close to the first page, last pagination item is a
link to the last page.
If we are close to the last page, first pagination item is a
link to the first page.
If we are away from both of first and last pages, pagination shows
links to the first page and to the last one.
I increased max amount of pagination items up. If numbers in items
consist of two or more digits pagination can overflow viewport
width on mobile devices, so I added styles to handle this situation.
Now if pagination overflows, items go to the new line.
Also code in the file is separated on few functions presently,
looks greater then it had looked before.